[refactor] Rename config.ts to app-config.ts
